HEEDLS-570 move update completion date function to progress service
