+++ DigitalLearningSolutions.Data/Services/ProgressService.cs
@@ -1,5 +1,6 @@
 ﻿namespace DigitalLearningSolutions.Data.Services
 {
+    using System;
     using System.Transactions;
     using DigitalLearningSolutions.Data.DataServices;
     using DigitalLearningSolutions.Data.Exceptions;
@@ -7,6 +8,7 @@
     public interface IProgressService
     {
         void UpdateSupervisor(int progressId, int? newSupervisorId);
+        void UpdateCompletionDate(int progressId, DateTime? completionDate);
     }
 
     public class ProgressService : IProgressService
@@ -47,6 +49,18 @@
             progressDataService.ClearAspProgressVerificationRequest(progressId);
 
             transaction.Complete();
+        }
+
+        public void UpdateCompletionDate(int progressId, DateTime? date)
+        {
+            var courseInfo = courseDataService.GetDelegateCourseInfoByProgressId(progressId);
+
+            if (courseInfo == null)
+            {
+                throw new ProgressNotFoundException($"No progress record found for ProgressID {progressId}");
+            }
+
+            courseDataService.SetCompletionDate(progressId, date);
         }
     }
 }
